refactor(dashboard): extract shared headers-editor component
Lifts the remote MCP header editor into a shared components/headers-editor.tsx
(draft-based editing, local validation, secret round-trip, Authorization always
excluded, optional catalog-suggestions prop) and re-points the remote MCP
HeadersSection at it. No behavior change.

Phase 3 of tunneled-MCP header support decomposition (source: 2ec86b2e9).
